refactor utils > ploting, adherence modules
