+++ app/adherence.py
@@ -1,9 +1,15 @@
+"""
+Module for monitoring patient medication adherence.
+"""
+import io
+import pickle
 from collections import defaultdict
-from datetime import datetime, timedelta
-import math
+from datetime import timedelta
+
 import numpy as np
-import plotly
-import plotly.graph_objects as go
+
+from modeling.preprocessing import preprocess
+
 
 # TODO: add more options
 DOSAGE_TO_FREQ = {
@@ -16,14 +22,45 @@ DOSAGE_TO_FREQ = {
 DAY_STD = {"day": 1, "week": 7, "month": 30}
 
 
+def adherence_model(data):
+    """Predict the number of pills remaining given a string of sensor data."""
+    classifier_path = "modeling/models/classifier.pkl"
+    regressor_path = "modeling/models/regressor.pkl"
+
+    # run the pill classifier process
+    X = preprocess(io.StringIO(data), regression=False)
+    classifier = pickle.load(open(classifier_path, "rb"))
+    classifier_pred = classifier.predict(X).item()
+
+    if classifier_pred == 0:
+        pred_string = "It does not appear you took any medication."
+        return {
+            "pred_string": pred_string,
+            "pred_type": "classification",
+            "pred": classifier_pred,
+        }
+    else:
+        # run the regression process
+        X = preprocess(io.StringIO(data), regression=True)
+
+        regressor = pickle.load(open(regressor_path, "rb"))
+        predicted_pills = regressor.predict(X).round().item()
+        predicted_pills = max(min(predicted_pills, 30), 1)
+
+        pred_string = f"It looks like you have {predicted_pills - 1} pills remaining."
+        return {
+            "pred_string": pred_string,
+            "pred_type": "regression",
+            "pred": predicted_pills,
+        }
+
+
 # TODO: Refactor such that the functions below can go in Prescription class.
-
-
 def get_next_refill_date(
     last_refill_date, duration, duration_unit, refills, refill_num
 ):
     """
-    Return next refill date based on most recent (last) refill date 
+    Return next refill date based on most recent (last) refill date
     and dosage information.
 
     last_refill_date: datetime
@@ -35,7 +72,7 @@ def get_next_refill_date(
 
     if refill_num == refills or refills == 0:
         return None
-    days_per_cycle = math.floor(duration * DAY_STD[duration_unit] / (refills + 1))
+    days_per_cycle = np.floor(duration * DAY_STD[duration_unit] / (refills + 1))
     return last_refill_date + timedelta(days=days_per_cycle)
 
 
@@ -108,105 +145,3 @@ def most_adhered_by_drug_name(prescriptions, n=5, measure="general"):
         for name, frac in sorted(med_adh.items(), key=lambda x: x[1], reverse=True)
     ]
     return sorted_adh[:n]
-
-
-def plot_top_general_adherence_by_drug_name(prescriptions, n=5):
-    top_general = most_adhered_by_drug_name(prescriptions, n=n, measure="general")
-
-    top_general_names = []
-    top_general_frac = []
-    for drug in top_general:
-        top_general_names.append(drug[0])
-        top_general_frac.append(drug[1])
-
-    fig = go.Figure(
-        data=[
-            go.Bar(name="General adherence", x=top_general_names, y=top_general_frac),
-        ]
-    )
-
-    fig.update_layout(
-        title="Top adhered medications",
-        xaxis_title="Date",
-        yaxis_title="Adherence %",
-        font=dict(color="#7f7f7f"),
-    )
-
-    output = plotly.offline.plot(fig, include_plotlyjs=False, output_type="div")
-    return output
-
-
-def plot_top_ontime_adherence_by_drug_name(prescriptions, n=5):
-    top_ontime = most_adhered_by_drug_name(prescriptions, n=n, measure="ontime")
-
-    top_ontime_names = []
-    top_ontime_frac = []
-    for drug in top_ontime:
-        top_ontime_names.append(drug[0])
-        top_ontime_frac.append(drug[1])
-
-    fig = go.Figure(
-        data=[go.Bar(name="ontime adherence", x=top_ontime_names, y=top_ontime_frac),]
-    )
-
-    fig.update_layout(
-        title="Top on-time medications",
-        xaxis_title="Date",
-        yaxis_title="On time %",
-        font=dict(color="#7f7f7f"),
-    )
-
-    output = plotly.offline.plot(fig, include_plotlyjs=False, output_type="div")
-    return output
-
-
-def plot_adherence_rates_over_time(patients, prescriptions):
-    """
-    Plot on-time and on-track adherence rates over time.
-    TODO: use real data
-    """
-
-    curr_date = datetime(year=2020, month=3, day=15)
-    end = datetime.now()
-    dates = []
-    while curr_date <= end:
-        curr_date += timedelta(days=1)
-        dates.append(curr_date)
-
-    np.random.seed(3)
-    x = np.arange(1, len(dates) + 1)
-    noise = np.random.uniform(-0.1, 0.1, len(dates))
-    patient_adh = np.clip(1 / (1 + np.exp(-(-0.3 + 0.03 * x))) + noise, 0.3, 0.96)
-    noise = np.random.uniform(-0.2, 0.05, len(dates))
-    rx_adh = patient_adh + noise
-
-    # Create traces
-    fig = go.Figure()
-    fig.add_trace(go.Scatter(x=dates, y=patient_adh, mode="lines", name="General"))
-    fig.add_trace(go.Scatter(x=dates, y=rx_adh, mode="lines", name="On-time"))
-
-    fig.update_layout(
-        title="Adherence rates over time",
-        xaxis_title="Date",
-        yaxis_title="Adherence %",
-        font=dict(color="#7f7f7f"),
-    )
-
-    fig.update_xaxes(
-        rangeslider_visible=True,
-        rangeselector=dict(
-            buttons=list(
-                [
-                    dict(count=1, label="1m", step="month", stepmode="backward"),
-                    dict(count=6, label="6m", step="month", stepmode="backward"),
-                    dict(count=1, label="YTD", step="year", stepmode="todate"),
-                    dict(count=1, label="1y", step="year", stepmode="backward"),
-                    dict(step="all"),
-                ]
-            )
-        ),
-    )
-
-    output = plotly.offline.plot(fig, include_plotlyjs=False, output_type="div")
-
-    return output
