+++ app/dashboard/views.py
@@ -3,8 +3,9 @@ from flask_login import current_user
 from flask_login import login_required
 
 from . import bp
+from app import adherence
+from app import ploting
 from app.models.persons import Patient
-from app.utils import *
 
 
 @bp.route("/dashboard")
@@ -15,9 +16,9 @@ def dashboard():
     rxs = []
     for p in patients:
         rxs += p.prescriptions
-    adh_over_time = plot_adherence_rates_over_time(patients, rxs)
-    top_general_adh = plot_top_general_adherence_by_drug_name(rxs, n=5)
-    top_ontime_adh = plot_top_ontime_adherence_by_drug_name(rxs, n=5)
+    adh_over_time = ploting.plot_adherence_rates_over_time(patients, rxs)
+    top_general_adh = ploting.plot_top_general_adherence_by_drug_name(rxs, n=5)
+    top_ontime_adh = ploting.plot_top_ontime_adherence_by_drug_name(rxs, n=5)
 
     return render_template(
         "dashboard.html",
